feat: 기존 auth -> member -> auth 패키지 참조 구조에서 auth -> member 단방향으로 바꿈
